+++ packages/@eventual/core/src/entity/entity.ts
@@ -5,7 +5,7 @@ import {
   EventualCallKind,
 } from "../internal/calls.js";
 import { getEntityHook } from "../internal/entity-hook.js";
-import { computeKeyDefinition } from "../internal/entity.js";
+import { computeKeyDefinition, KeyDefinition } from "../internal/entity.js";
 import { entities } from "../internal/global.js";
 import {
   EntityIndexSpec,
@@ -35,13 +35,19 @@ export type AttributeBinaryValue =
   | BigInt64Array
   | BigUint64Array;
 
-export type AttributeValue =
-  | Attributes
+type AttributeScalarValue =
+  | null
+  | undefined
+  | bigint
   | string
   | number
   | boolean
-  | AttributeBinaryValue
-  | Set<string | number | boolean | AttributeBinaryValue>
+  | AttributeBinaryValue;
+
+export type AttributeValue =
+  | Attributes
+  | AttributeScalarValue
+  | Set<AttributeScalarValue>
   | AttributeValue[];
 
 export interface Attributes {
@@ -71,7 +77,8 @@ export interface Entity<
     "attributes" | "streams" | "partition" | "sort" | "indices"
   > {
   kind: "Entity";
-  attributes: z.ZodObject<EntityZodShape<Attr>>;
+  key: KeyDefinition;
+  attributes: ZodAttributesObject<Attr>;
   indices: EntityIndex[];
   streams: EntityStream<Attr, Partition, Sort>[];
   /**
@@ -145,12 +152,24 @@ export const Entity = {
   },
 };
 
+/**
+ * Tries the {@link Attributes} type to the computed output of the object.
+ *
+ * TODO: extend this type to support intersection and union.
+ */
+export type ZodAttributesObject<T extends Attributes> = z.ZodObject<
+  any,
+  any,
+  any,
+  T
+>;
+
 export interface EntityOptions<
   Attr extends Attributes,
   Partition extends CompositeKeyPart<Attr>,
   Sort extends CompositeKeyPart<Attr> | undefined = undefined
 > {
-  attributes: z.ZodObject<EntityZodShape<Attr>> | EntityZodShape<Attr>;
+  attributes: ZodAttributesObject<Attr> | EntityZodShape<Attr>;
   partition: Partition;
   sort?: Sort;
 }
@@ -238,7 +257,7 @@ export function entity<
   const attributes =
     options.attributes instanceof z.ZodObject
       ? options.attributes
-      : z.object(options.attributes);
+      : (z.object(options.attributes) as unknown as ZodAttributesObject<Attr>);
 
   const entity: Entity<Attr, Partition, Sort> = {
     // @ts-ignore