+++ packages/@eventual/aws-cdk/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.37.2 (2023-05-14)
+
+**Note:** Version bump only for package @eventual/aws-cdk
+
+## 0.37.1 (2023-05-14)
+
+**Note:** Version bump only for package @eventual/aws-cdk
+
 # 0.37.0 (2023-05-12)
 
 ### Features